--- src/features/post-place/types/city-type.ts
@@ -1,4 +1,0 @@
-export interface City {
-  name: string;
-  id: string;
-}
